added/fixed tests
