+++ components/src/components/molecules/Modal/Modal.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 
 import { ThemeProvider } from 'styled-components'
 
-import { cleanup, render, screen, userEvent } from '@/test'
+import { cleanup, render, screen, waitFor } from '@/test'
 
 import { Modal } from './Modal'
 import { lightTheme } from '@/src/tokens'
@@ -16,7 +16,9 @@ describe('<Modal />', () => {
         <Modal open>Modal</Modal>
       </ThemeProvider>,
     )
-    expect(screen.getByText('Modal')).toBeVisible()
+    waitFor(() => expect(screen.getByText('Modal')).toBeVisible(), {
+      timeout: 300,
+    })
   })
 
   it('should not be visible if not open', () => {
@@ -25,31 +27,8 @@ describe('<Modal />', () => {
         <Modal open={false}>Modal</Modal>
       </ThemeProvider>,
     )
-    expect(screen.queryByText('Modal')).toBeNull()
-  })
-
-  it('should display close icon if callback is provided', () => {
-    const mockCallback = jest.fn()
-    render(
-      <ThemeProvider theme={lightTheme}>
-        <Modal open onDismiss={mockCallback}>
-          Modal
-        </Modal>
-      </ThemeProvider>,
-    )
-    expect(screen.getByTestId('close-icon')).toBeVisible()
-  })
-
-  it('should call callback if close icon is clicked', () => {
-    const mockCallback = jest.fn()
-    render(
-      <ThemeProvider theme={lightTheme}>
-        <Modal open onDismiss={mockCallback}>
-          Modal
-        </Modal>
-      </ThemeProvider>,
-    )
-    userEvent.click(screen.getByTestId('close-icon'))
-    expect(mockCallback).toHaveBeenCalled()
+    waitFor(() => expect(screen.getByText('Modal')).toBeNull(), {
+      timeout: 300,
+    })
   })
 })
